requirements.txt
python-telegram-bot==21.7
aiohttp==3.9.5
aiofiles==23.2.1
python-dotenv==1.0.1
pytz==2024.1
# Дополнительные зависимости для оптимизации
cryptography>=41.0.0
asyncio-throttle>=1.0.0
# Планировщик задач для сброса счетчиков
APScheduler==3.10.4
# Зависимости для распознавания голоса (временно отключены)
# SpeechRecognition==3.10.0
# pydub==0.25.1
# Зависимости для определения часового пояса по координатам
timezonefinder==6.2.0

aiosqlite==0.20.0

pytest==8.3.2
pytest-asyncio==0.23.8

# Дополнительные зависимости для продакшена
gunicorn==21.2.0
psutil==5.9.8
